refactor: namespace

+++ include/conv/conv_minplus_cv.hpp
@@ -17,7 +17,7 @@ CEXP vec<T> conv_minplus_cv(vec<T> CR a, vec<T> CR b) NE {
         i32 i1 = (i32)k - (i32)j1, i2 = (i32)k - (i32)j2;
         if (i2 < 0) return 1;
         if (i1 >= (i32)n) return 0;
-        return a[i1] + b[j1] < a[i2] + b[j2];
+        return a[(u32)i1] + b[j1] < a[(u32)i2] + b[j2];
       });
   vec<T> c(n + m - 1);
   flt_ (u32, k, 0, n + m - 1) c[k] = a[k - argmin[k]] + b[argmin[k]];
